Refactor test folders
