+++ package.json
@@ -16,15 +16,16 @@
     "@mui/icons-material": "^5.15.10",
     "@mui/joy": "^5.0.0-beta.28",
     "@mui/material": "^5.15.10",
-    "@testing-library/react": "^14.2.1",
     "axios": "^1.6.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.22.1",
-    "vitest": "^1.3.1"
+    "react-router-dom": "^6.22.1"
   },
   "devDependencies": {
-    "@types/react": "^18.2.55",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1",
+    "@types/jest": "^29.5.12",
+    "@types/react": "^18.2.60",
     "@types/react-dom": "^18.2.19",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
@@ -34,6 +35,7 @@
     "eslint-plugin-react-refresh": "^0.4.5",
     "jsdom": "^24.0.0",
     "typescript": "^5.2.2",
-    "vite": "^5.1.1"
+    "vite": "^5.1.1",
+    "vitest": "^1.3.1"
   }
 }
